GP2-657: Merge both templatetag test files into one
Still keeps them as a mix of pytest and unittest, but at least they are in a single place for discovery
